+++ src/main/java/com/prgrms/needit/domain/board/activity/entity/ActivityImage.java
@@ -1,6 +1,6 @@
 package com.prgrms.needit.domain.board.activity.entity;
 
-import com.prgrms.needit.common.domain.BaseEntity;
+import com.prgrms.needit.common.domain.entity.BaseEntity;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.JoinColumn;
