+++ src/main/java/com/prgrms/needit/domain/board/donation/repository/ThemeTagRepository.java
@@ -1,6 +1,6 @@
 package com.prgrms.needit.domain.board.donation.repository;
 
-import com.prgrms.needit.common.domain.ThemeTag;
+import com.prgrms.needit.common.domain.entity.ThemeTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ThemeTagRepository extends JpaRepository<ThemeTag, Long> {
