+++ src/main/java/com/prgrms/needit/domain/center/entity/Center.java
@@ -1,6 +1,6 @@
 package com.prgrms.needit.domain.center.entity;
 
-import com.prgrms.needit.common.domain.BaseEntity;
+import com.prgrms.needit.common.domain.entity.BaseEntity;
 import com.prgrms.needit.common.enums.UserType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
