+++ src/main/java/com/prgrms/needit/domain/message/entity/MessageContract.java
@@ -1,6 +1,6 @@
 package com.prgrms.needit.domain.message.entity;
 
-import com.prgrms.needit.common.domain.BaseEntity;
+import com.prgrms.needit.common.domain.entity.BaseEntity;
 import com.prgrms.needit.common.enums.UserType;
 import java.time.LocalDateTime;
 import javax.persistence.Column;
